Added plugins for different versions
