+++ src/Persistence/Initialization/Updates/RegenerationsRefactorPlugIn075.cs
@@ -1,0 +1,30 @@
+// <copyright file="RegenerationsRefactorPlugIn075.cs" company="MUnique">
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// </copyright>
+
+namespace MUnique.OpenMU.Persistence.Initialization.Updates;
+
+using System.Runtime.InteropServices;
+using MUnique.OpenMU.DataModel.Configuration;
+using MUnique.OpenMU.PlugIns;
+
+/// <summary>
+/// This update fixes and reworks some regeneration attributes (health, mana, ability).
+/// </summary>
+[PlugIn]
+[Display(Name = PlugInName, Description = PlugInDescription)]
+[Guid("F079402E-F557-423C-B376-80A1B87D842D")]
+public class RegenerationsRefactorPlugIn075 : RegenerationsRefactorPlugInBase
+{
+    /// <inheritdoc />
+    public override UpdateVersion Version => UpdateVersion.RegenerationsRefactor075;
+
+    /// <inheritdoc />
+    public override string DataInitializationKey => Version075.DataInitialization.Id;
+
+    /// <inheritdoc />
+    protected override async ValueTask ApplyAsync(IContext context, GameConfiguration gameConfiguration)
+    {
+        await base.ApplyAsync(context, gameConfiguration).ConfigureAwait(false);
+    }
+}
